将chat改为伪流式输出

+++ backend/src/main/java/com/example/backend/vo/MessageVO.java
@@ -16,7 +16,7 @@ public class MessageVO {
     private Long videoId;
     private String url;
     private Integer part;
-    private Long subtitleCount;
+    private Integer subtitleCount;
     private List<Message> messages;
 
 }
